CRUD ok

+++ services/officialLeague - Connexion/src/main/java/ch/heigvd/amt/repositories/UserRepository.java
@@ -3,5 +3,5 @@ package ch.heigvd.amt.repositories;
 import ch.heigvd.amt.entities.UserEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserEntity, Long> {
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
 }
